+++ internals-js/src/utils.ts
@@ -388,3 +388,9 @@ export function printHumanReadableList(
 export type Concrete<Type> = {
   [Property in keyof Type]-?: Type[Property];
 };
+
+// for use with Array.filter
+// Example:
+//   const x = [1,2,undefined];
+//   const y: number[] = x.filter(isDefined);
+export const isDefined = <T>(t: T | undefined): t is T => t === undefined ? false : true;