+++ composition-js/CHANGELOG.md
@@ -2,7 +2,12 @@
 
 This CHANGELOG pertains only to Apollo Federation packages in the 2.x range. The Federation v0.x equivalent for this package can be found [here](https://github.com/apollographql/federation/blob/version-0.x/federation-js/CHANGELOG.md) on the `version-0.x` branch of this repo.
 
-- Warn on merging inconsistent non-repeatable directive applications instead of failing composition [#1840](https://github.com/apollographql/federation/pull/1840).
+## 2.1.0-alpha.2
+- Add `@composeDirective` directive to specify directives that should be merged to the supergraph during composition [PR #1996](https://github.com/apollographql/federation/pull/1996).
+
+## 2.1.0-alpha.1
+
+- Warn on merging inconsistent non-repeatable directive applications instead of failing composition [PR #1840](https://github.com/apollographql/federation/pull/1840).
 
 ## 2.1.0-alpha.0